Refactored value_mappings and Factor base class; improved .rvs() speed.

+++ graphmodels/multinomial/multinomial.py
@@ -7,7 +7,7 @@ def multinomial(p, size=1):
     assert (1 <= p.ndim <= 2)
     if p.ndim == 1:
         p = np.array([p] * size)
-    res = np.zeros(p.shape[0])
+    res = np.zeros(p.shape[0]).astype('int32')
     cs = np.cumsum(p, axis=1).astype('float64')
     gen_multinomial(res, cs)
     return res
